+++ tests/cli-json-input.test.ts
@@ -1,6 +1,6 @@
 import { runAndCatch } from '@carnesen/run-and-catch';
-import { CLI_USAGE_ERROR } from './cli-usage-error';
-import { CliJsonInput } from './cli-json-input';
+import { CliJsonInput } from '../src/cli-json-input';
+import { CLI_USAGE_ERROR } from '../src/cli-usage-error';
 
 const description = 'foo bar baz';
 const hidden = true;
