+++ tests/examples/hidden-option.test.ts
@@ -1,5 +1,5 @@
-import { root } from '../hidden-option';
-import { CliArgvInterface } from '../../cli-argv-interface';
+import { CliArgvInterface } from '../../src/cli-argv-interface';
+import { root } from '../../src/examples/hidden-option';
 
 const cli = CliArgvInterface(root);
