+++ tests/find-version.test.ts
@@ -1,4 +1,4 @@
-import { findVersion } from './find-version';
+import { findVersion } from '../src/find-version';
 
 describe(findVersion.name, () => {
   it('finds a package version by looking upwards in the filesystem from require.main', async () => {
