+++ tests/last-command.test.ts
@@ -1,6 +1,6 @@
-import { CliBranch } from './cli-branch';
-import { CliLeaf } from './cli-leaf';
-import { LastCommand } from './last-command';
+import { CliBranch } from '../src/cli-branch';
+import { CliLeaf } from '../src/cli-leaf';
+import { LastCommand } from '../src/last-command';
 
 const leaf = CliLeaf({
   name: 'l',
