+++ tsconfig.eslint.json
@@ -1,5 +1,5 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./src", "./tests", "./tests-integration"],
+  "include": ["./src", "./tests"],
   "compilerOptions": {}
 }
