+++ tsconfig.json
@@ -1,10 +1,88 @@
 {
-  "extends": "@alwaysai/tsconfig/nodejs/tsconfig.json",
-  "include": [
-    "src"
-  ],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./lib"
+    "extends": "@alwaysai/tsconfig/nodejs/tsconfig.json",
+    "include": ["src"],
+    "compilerOptions": {
+      "rootDir": "./src",
+      "outDir": "./lib",
+      // "allowJs": false,
+      "allowSyntheticDefaultImports": true,
+      // "allowUnreachableCode": false,
+      // "allowUnusedLabels": false,
+      "alwaysStrict": false,
+      // "baseUrl": ".",
+      // "build": false,
+      // "charset": "utf8",
+      // "checkJs": false,
+      // "composite": true,
+      "declaration": true,
+      // "declarationDir": ".",
+      "declarationMap": true,
+      // "diagnostics": false,
+      // "disableSizeLimit": false,
+      "downlevelIteration": true,
+      // "emitBOM": false,
+      // "emitDeclarationOnly": false,
+      // "emitDecoratorMetadata": false,
+      // "esModuleInterop": false,
+      // "experimentalDecorators": false,
+      "forceConsistentCasingInFileNames": true,
+      // "importHelpers": false,
+      // "inlineSourceMap": false,
+      // "inlineSources": false,
+      // "isolatedModules": false,
+      // "jsx": "preserve",
+      // "jsxFactory": "React.createElement",
+      // "keyofStringsOnly": false,
+      // "lib": [],
+      // "listEmittedFiles": false,
+      // "listFiles": false,
+      // "locale": "en",
+      // "mapRoot": ".",
+      // "maxNodeModuleJsDepth": 0,
+      "module": "commonjs",
+      "moduleResolution": "node",
+      "newLine": "lf",
+      // "noEmit": false,
+      // "noEmitHelpers": false,
+      "noEmitOnError": false,
+      "noErrorTruncation": true,
+      // "noFallthroughCasesInSwitch": false,
+      "noImplicitAny": false,
+      "noImplicitReturns": false,
+      "noImplicitThis": false,
+      // "noImplicitUseStrict": false,
+      // "noLib": false,
+      // "noStrictGenericChecks": false,
+      // "noResolve": false,
+      "noUnusedLocals": true,
+      "noUnusedParameters": true,
+      // "out": DEPRECATED
+      // "outDir": "lib",
+      // "outFile": "dist/lib.js",
+      // "paths": [],
+      // "preserveConstEnums": false,
+      // "preserveSymlinks": false,
+      // "preserveWatchOutput": false,
+      // "pretty": true,
+      // "reactNamespace": DEPRECATED
+      // "removeComments": false,
+      "resolveJsonModule": true,
+      // "rootDir": "src",
+      // "rootDirs": ["src"],
+      // "skipDefaultLibCheck": DEPRECATED
+      // "skipLibCheck": false,
+      "sourceMap": true,
+      // "sourceRoot": ".",
+      "strict": false,
+      "strictFunctionTypes": false,
+      "strictNullChecks": true,
+      "strictPropertyInitialization": false
+      // "stripInternal": false,
+      // "suppressExcessPropertyErrors": false,
+      // "suppressImplicitAnyIndexErrors": false,
+      // "target": "esnext",
+      // "traceResolution": false,
+      // "types": [],
+      // "typeRoots": [],
+    }
   }
-}
